[bphh-531] Refactor requirement block displays to accordion for template builder
